refactor: splits index file into multiple parts
moved main bundling loop out of index file into separate api
