+++ test/bundle-test.ts
@@ -1,0 +1,28 @@
+import test from 'ava';
+import { HostMock, virtualModule } from './helper';
+
+import { bundle } from '../src/bundle';
+
+test('bundle should bundle the given entry-point and its dependencies', t => {
+  const host = new HostMock({
+    'entry-point.js': `
+      import fn from './dependency';
+      fn();
+    `,
+    './dependency.js': `
+      export default function() {
+        callme();
+      }
+    `
+  });
+
+  const bundled = bundle('entry-point.js', host);
+
+  let called = false;
+  virtualModule(bundled, {
+    callme: function(): void {
+      called = true;
+    }
+  });
+  t.true(called);
+});
